Cambio en el nombre de la carpeta ciudadano, cambio en el nombre de los archivos ciudadanos y auth, modificacion del codigo, login del ciudadano funcionando
